% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{add_stats}
\alias{add_stats}
\title{Add test statistics to output}
\usage{
add_stats(tab, ...)
}
\arguments{
\item{tab}{A \code{cross_table} object.}

\item{...}{Further arguments, currently passed down to
\code{\link{fisher.test}}.}
}
\description{
Currently this function computes a \code{\link{fisher.test}} on the data and
displays the resulting output next to the crosstab.
}
\examples{
titanic \%>\%
  crosstab(Survived ~ Sex) \%>\%
  add_stats()
}

