% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tab.R
\name{build_tab}
\alias{build_tab}
\title{Build a table}
\usage{
build_tab(tab)
}
\arguments{
\item{tab}{A tab object}
}
\value{
Returns a matrix with counts and proportions, seperated by "<br>".
   Currently just a single two dimensional table is created.
}
\description{
Builds a \code{matrix} for use with \code{\link[htmlTable]{htmlTable}}.
}
\keyword{internal}

