% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tab.R
\name{compute_frequencies}
\alias{compute_frequencies}
\title{Compute counts and frequencies}
\usage{
compute_frequencies(x)
}
\arguments{
\item{x}{A tab}
}
\value{
A \code{data.frame} with counts and proportions of for dependent and
independent variables.
}
\description{
Computes counts and frequencies from a \code{\link{tab}}.
}
\details{
Unused factor levels are retained
}
\keyword{internal}

