% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{create_page}
\alias{create_page}
\title{HTML parts of page}
\usage{
create_page(html_table)
}
\arguments{
\item{html_table}{a bare HTML table, created with \code{htmlTable}.}
}
\value{
a \code{tagList} with registered dependencies.
}
\description{
This function currently serves to prepare the pagelayout.
}
\details{
Later on this function should be more flexible insofar as the content of the
page should be dependent on the content of the table.
}

