% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_table.R
\name{cross_table}
\alias{cross_table}
\title{cross_table}
\usage{
cross_table(data, x)
}
\arguments{
\item{data}{\code{data.frame} which contains the data, given in \code{x}}

\item{x}{A formula specifying \code{dependent ~ independent}.

Strictly speaking there is no dependent or independet variable in
cross-tabulation. Nevertheless it is common to calculate percentages, for
which the distinction between dependent and independent variable is
meaningful. \code{cross_table} lets you specify the model in a memorizable
way.}
}
\description{
build cross table from formula
}

