% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{crosstab}
\usage{
crosstab(data = NULL, x = NULL, layer = NULL)
}
\arguments{
\item{data}{\code{data.frame} which contains the data, given in \code{x}}

\item{x}{A formula specifying \code{dependent ~ independent}.}

\item{layer}{A third variable which is used to construct a three-dimensional
   crosstab.

Strictly speaking there is no dependent or independet variable in
cross-tabulation. Nevertheless it is common to calculate percentages, for
which the distinction between dependent and independent variable is
meaningful. \code{crosstab} lets you specify the model in a memorizable
way.}
}
\description{
build cross table from formula
}
\examples{
titanic \%>\%
  crosstab(Survived ~ Sex)
}

