% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_factor}
\alias{expand_factor}
\title{Expand factor}
\usage{
expand_factor(data, x)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{x}{a quoted name of the variable within the \code{data.frame}, which
should be expanded}
}
\value{
A \code{data.frame} with one named variable, containing all levels
   of x.
}
\description{
\code{expand_factor} expands a factor to include all levels, even those,
which were not observed in the data. It returns a \code{data.frame} which
can subsequently used to condunct a \code{left_join}.
}

