% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_twoside_formula}
\alias{is_twoside_formula}
\title{Is a formula?}
\usage{
is_twoside_formula(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\description{
Checks if input is a formula.
}
\examples{
x <- disp ~ am
crosstabr:::is_twoside_formula(x)
}
\keyword{internal}

