% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{prepare_table}
\alias{prepare_table}
\title{Creates a HTML table}
\usage{
prepare_table(x)
}
\arguments{
\item{x}{A matrix, created by \code{build_tab}.}
}
\value{
A table in HTML format, without inline styling.
}
\description{
Creates a HTML table
}
\keyword{internal}

