% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cross_table}
\alias{print.cross_table}
\title{Print method for cross_table}
\usage{
\method{print}{cross_table}(tab, ...)
}
\arguments{
\item{tab}{A cross_table object.}
}
\description{
The cross tabulation is rendered as a html, which can be viewed in RStudio's
viewer pane.
}
\keyword{internal}

