% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.crosstab}
\alias{print.crosstab}
\title{Print method for crosstab}
\usage{
\method{print}{crosstab}(x, ...)
}
\arguments{
\item{x}{A crosstab object.}
}
\description{
The cross tabulation is rendered as a html, which can be viewed in RStudio's
viewer pane.
}
\keyword{internal}

