% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{print.htest_custom}
\alias{print.htest_custom}
\title{Rewrite print.method for tests to exclude data.}
\usage{
\method{print}{htest_custom}(x, digits = getOption("digits"),
  prefix = "\\t", ...)
}
\arguments{
\item{x}{The result of a statistical test of dependence (chisq, fisher, ...)}
}
\description{
Rewrite print.method for tests to exclude data.
}
\keyword{internal}

