% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{tab}
\alias{tab}
\title{A tab-object}
\description{
The \code{tab} is the core of \code{crosstabr}.
}
\details{
The \code{tab} is an object of class \code{crosstab}. It contains the
formula of the \code{crosstab}, from which a \code{\link{model.frame}} was
built, as well as other arguments which determine the apperance of the final
cross table.
}
\keyword{internal}

