% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Data on Titanic survival.}
\format{A data frame with 2201 rows and 4 columns.
The levels are as follows:
\describe{
  \item{Class}{1st, 2nd, 3rd, Crew}
  \item{Sex}{Male, Female}
  \item{Age}{Child, Adult}
  \item{Survived}{No, Yes}
}}
\usage{
titanic
}
\description{
This dataset is a reshaped form of \link[datasets]{Titanic}.
}
\keyword{datasets}

