% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{add_stats}
\alias{add_stats}
\title{Add test statistics to output}
\usage{
add_stats(tab, ...)
}
\arguments{
\item{tab}{A \code{cross_table} object.}

\item{...}{Names of statistical tests to perform on the data, i.e.
\code{\link{fisher.test}} or \code{\link{chisq.test}}. The function must
accept a \code{table} as input an can be an anonymous function.
If no test is supplied, \code{\link[vcd]{assocstats}} is performed as
default.}
}
\description{
Currently this function computes a \code{\link{fisher.test}} on the data and
displays the resulting output next to the crosstab.
}
\examples{
titanic \%>\%
  crosstab(Survived ~ Sex) \%>\%
  add_stats()
}

